feat: adding env url
