+++ .gitignore
@@ -1,2 +1,1 @@
-/httpclient
-/ginclient
+/client
