--- cmd/ginclient/client.go
@@ -1,20 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"io"
-	"net/http"
-	"time"
-)
-
-func main() {
-	c := http.Client{Timeout: time.Duration(1) * time.Second}
-	response, err := c.Get("http://localhost:8089/datetime")
-	if err != nil {
-		fmt.Printf("Error %s", err)
-		return
-	}
-	defer response.Body.Close()
-	body, _ := io.ReadAll(response.Body)
-	fmt.Printf("Body : %s", body)
-}
